Split resolver type metadata default tests
